json parser

+++ src/main/java/com/asterix/weatherapp/dto/ForcastDay/Day.java
@@ -3,41 +3,41 @@ package com.asterix.weatherapp.dto.ForcastDay;
 import com.asterix.weatherapp.dto.Condition;
 
 public class Day {
-    private Integer maxtemp_c;
-    private Integer mintemp_c;
-    private Integer avghumidity;
+    private Double maxtemp_c;
+    private Double mintemp_c;
+    private Double avghumidity;
     private Condition condition;
 
     public Day(){}
 
-    public Day(Integer maxtemp_c, Integer mintemp_c, Integer avghumidity, Condition condition) {
+    public Day(Double maxtemp_c, Double mintemp_c, Double avghumidity, Condition condition) {
         this.maxtemp_c = maxtemp_c;
         this.mintemp_c = mintemp_c;
         this.avghumidity = avghumidity;
         this.condition = condition;
     }
 
-    public Integer getMaxtemp_c() {
+    public Double getMaxtemp_c() {
         return maxtemp_c;
     }
 
-    public void setMaxtemp_c(Integer maxtemp_c) {
+    public void setMaxtemp_c(Double maxtemp_c) {
         this.maxtemp_c = maxtemp_c;
     }
 
-    public Integer getMintemp_c() {
+    public Double getMintemp_c() {
         return mintemp_c;
     }
 
-    public void setMintemp_c(Integer mintemp_c) {
+    public void setMintemp_c(Double mintemp_c) {
         this.mintemp_c = mintemp_c;
     }
 
-    public Integer getAvghumidity() {
+    public Double getAvghumidity() {
         return avghumidity;
     }
 
-    public void setAvghumidity(Integer avghumidity) {
+    public void setAvghumidity(Double avghumidity) {
         this.avghumidity = avghumidity;
     }
